# Add -mlongcalls to the Xtensa compiler flags.