--- a/tensorflow/lite/micro/tools/make/targets/xtensa_makefile.inc
+++ b/tensorflow/lite/micro/tools/make/targets/xtensa_makefile.inc
@@ -39,7 +39,8 @@ PLATFORM_FLAGS = \
   -mcoproc \
   -DMAX_RFFT_PWR=9 \
   -DMIN_RFFT_PWR=MAX_RFFT_PWR \
-  $(TARGET_ARCH_DEFINES)
+  $(TARGET_ARCH_DEFINES) \
+  -mlongcalls
 
 ifeq ($(BUILD_TYPE), release)
   PLATFORM_FLAGS += -Wno-unused-private-field